Modernize package metadata + structure
* Converted setup.cfg etc. -> pyproject.toml
* Moved the plugin itself from `__init__` to a `plugin` submodule,
  matching our plugin template for Sopel 8.x
* Populated the NEWS file, ready to create new releases
