+++ pyproject.toml
@@ -1,0 +1,55 @@
+[build-system]
+requires = ["setuptools>=68.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+platforms = ["Linux x86, x86-64"]
+
+[tool.setuptools.packages.find]
+include = ["sopel_bugzilla", "sopel_bugzilla.*"]
+namespaces = false
+
+[tool.setuptools.dynamic]
+readme = { file=["README.md", "NEWS"], content-type="text/markdown" }
+
+[project]
+name = "sopel-bugzilla"
+version = "0.1.0"
+description = "Domain-based Bugzilla link handler plugin for Sopel IRC bots."
+
+authors = [
+  { name = "Ari Koivula" },
+  { name = "Elad Alfassa" },
+  { name = "Embolalia" },
+]
+maintainers = [
+  { name="dgw", email="dgw@technobabbl.es" },
+]
+
+license = "EFL-2.0"
+license-files = ["COPYING"]
+dynamic = ["readme"]
+
+classifiers = [
+  "Intended Audience :: Developers",
+  "Intended Audience :: System Administrators",
+  "Topic :: Communications :: Chat :: Internet Relay Chat",
+]
+keywords = [
+  "sopel",
+  "plugin",
+  "bot",
+  "irc",
+]
+
+requires-python = ">=3.8, <4"
+dependencies = [
+  "sopel>=8.0",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/sopel-irc/sopel-bugzilla"
+"Bug Tracker" = "https://github.com/sopel-irc/sopel-bugzilla/issues"
+
+[project.entry-points."sopel.plugins"]
+"bugzilla" = "sopel_bugzilla.plugin"
